Viimased lihvid teacherAdminPagele ning personViews emailile

+++ Pages/Autokool/Admin/TeachersAdminPage.cs
@@ -8,7 +8,7 @@ using System.Threading.Tasks;
 
 namespace Autokool.Pages.Autokool.Admin
 {
-    [Authorize(Roles = "Teacher, Administrator")]
+    [Authorize(Roles = "Administrator")]
     public class TeachersAdminPage : TeachersBasePage<TeachersAdminPage>
     {
         private readonly UserManager<ApplicationUser> _userManager;
@@ -19,7 +19,6 @@ namespace Autokool.Pages.Autokool.Admin
         {
             _userManager = userManager;
         }
-
         public override async Task<IActionResult> OnPostCreateAsync(string sortOrder, string searchString, int? pageIndex, string fixedFilter, string fixedValue)
         {
             if (!await addObject(sortOrder, searchString, pageIndex, fixedFilter, fixedValue)
@@ -32,7 +31,6 @@ namespace Autokool.Pages.Autokool.Admin
                 FirstName = Item.FirstName,
                 LastName = Item.Name
             };
-            UserId = user.Id;//vaata!!!
             var result = await _userManager.CreateAsync(user, user.FirstName.ToString() + user.LastName.ToString() + "1'");
             if (result.Succeeded)
             {
@@ -59,6 +57,10 @@ namespace Autokool.Pages.Autokool.Admin
                 FirstName = Item.FirstName,
                 LastName = Item.Name
             };
+            if (user == null)
+            {
+                return NotFound();
+            }
             var result = await _userManager.CreateAsync(user, user.FirstName.ToString() + user.LastName.ToString() + "1'");
             if (result.Succeeded)
             {
@@ -66,14 +68,14 @@ namespace Autokool.Pages.Autokool.Admin
             }
             return Redirect(IndexUrl.ToString());
         }
-        public override Task<IActionResult> OnGetDeleteAsync(string id, string sortOrder, string searchString, int? pageIndex, string fixedFilter, string fixedValue)
-        {
-            return base.OnGetDeleteAsync(id, sortOrder, searchString, pageIndex, fixedFilter, fixedValue);
-        }
         public override async Task<IActionResult> OnPostDeleteAsync(string id, string sortOrder, string searchString, int pageIndex, string fixedFilter, string fixedValue)
         {
             await deleteObject(id, sortOrder, searchString, pageIndex, fixedFilter, fixedValue).ConfigureAwait(true);
             var userToDelete = await _userManager.FindByIdAsync(id);
+            if (userToDelete == null)
+            {
+                return NotFound();
+            }
             await _userManager.DeleteAsync(userToDelete);
             return Redirect(IndexUrl.ToString());
         }
